Use JavaScriptCore as interpreter
This slipped through the cracks because of the presence of cscript
on Windows machines, but we can't rely on that OOTB due to Windows
vs. *NIX paths (not to mention that I wouldn't WANT to).

The preferred interpreter on *NIX systems is ruby-therubyracer, but
that requires v8, which BRs the quite-broken-on-Cygwin gyp.  The
alternatives are ruby-johnson (ancient bundled mozjs) and js(1) (which
is deprecated upstream per github issue #98).  Fixing the path of
jsc from the OSX default seems to work fine, but probably cannot be
upstreamed.